requirements.txt
streamlit
pymongo
pillow
